makeing sessions_records share expanded state so switching between mobile and desktop dose not reset state
